+++ src/mruby_integration/core/gestures.cpp
@@ -4,17 +4,6 @@
 #include "ruby/core/gestures.hpp"
 
 auto
-mrb_set_gestures_enabled(mrb_state* mrb, mrb_value) -> mrb_value
-{
-  mrb_int flags;
-  mrb_get_args(mrb, "i", &flags);
-
-  SetGesturesEnabled(flags);
-
-  return mrb_nil_value();
-}
-
-auto
 mrb_get_gesture_detected(mrb_state* mrb, mrb_value) -> mrb_value
 {
   return mrb_int_value(mrb, GetGestureDetected());
@@ -23,11 +12,6 @@ mrb_get_gesture_detected(mrb_state* mrb, mrb_value) -> mrb_value
 void
 append_core_gestures(mrb_state* mrb)
 {
-  mrb_define_method(mrb,
-                    mrb->kernel_module,
-                    "set_gestures_enabled",
-                    mrb_get_gesture_detected,
-                    MRB_ARGS_REQ(1));
   mrb_define_method(mrb,
                     mrb->kernel_module,
                     "get_gesture_detected",