+++ src/mruby_integration/models.cpp
@@ -5,6 +5,7 @@
 #include "mruby_integration/models/cursor.hpp"
 #include "mruby_integration/models/font.hpp"
 #include "mruby_integration/models/gamepad.hpp"
+#include "mruby_integration/models/gesture.hpp"
 #include "mruby_integration/models/image.hpp"
 #include "mruby_integration/models/key.hpp"
 #include "mruby_integration/models/line.hpp"
@@ -30,6 +31,7 @@ append_models(mrb_state* mrb)
   append_models_Cursor(mrb);
   append_models_Font(mrb);
   append_models_Gamepad(mrb);
+  append_models_Gesture(mrb);
   append_models_Image(mrb);
   append_models_Key(mrb);
   append_models_Line(mrb);